add team app and create Team model
