+++ ticket_sales/ticket_sales/settings.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = [
     'core',
     'events',
     'crispy_forms',
+    'team',
 ]
 
 MIDDLEWARE = [
@@ -84,13 +85,8 @@ WSGI_APPLICATION = 'ticket_sales.wsgi.application'
 
 DATABASES = {
     'default': {
-        'ENGINE': os.environ.get('APP_DB_ENGINE', 'django.db.backends.sqlite3'),
-        'NAME': os.environ.get('DB_NAME', 'db.sqlite'),
-        'USER': os.environ.get('DB_USER', ''),
-        'PASSWORD': os.environ.get('DB_PASSWORD', ''),
-        'HOST': os.environ.get('DB_HOST', None),
-        'PORT': os.environ.get('DB_PORT', None),
-        'CONN_MAX_AGE': 5,  # 5 sec
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': BASE_DIR / 'db.sqlite3',
     }
 }
